test(coding-agent): split provider timeout recovery coverage
